refactored styles and codes
