+++ src/tsconfig.json
@@ -19,13 +19,12 @@
         "./*"
       ]
     },
-    "resolveJsonModule": true // Added this line
+    "resolveJsonModule": true
   },
   "include": [
     "./**/*",
-    ".expo/types/**/*.ts",
-    "expo-env.d.ts"
-, "../app.config.js"  ],
+    "../app.config.js"
+  ],
   "exclude": [
     "node_modules"
   ],
